I changed the name of question1 code to "sms_pycode.py" for easy import. This commit includes first unittest
